finished up startgame backend on GameLobby activity

+++ client/WhereWolf2014/src/edu/utexas/LI/wherewolf/CreateGameRequest.java
@@ -45,7 +45,7 @@ public class CreateGameRequest extends BasicRequest{
 	{
 
 		JSONObject jResults = jObject.getJSONObject("results");
-		int gameID = jResults.getInt("game_id");
+		long gameID = (long) jResults.getInt("game_id");
 
 		return new CreateGameResponse("success",
 				"Successfully created the game",
