+++ client/WhereWolf2014/src/edu/utexas/LI/wherewolf/CreateGameResponse.java
@@ -2,20 +2,20 @@ package edu.utexas.LI.wherewolf;
 
 public final class CreateGameResponse extends BasicResponse {
 
-	protected int gameID;
+	protected long gameID;
 
 	public CreateGameResponse (String status, String errorMessage)
 	{
 		super(status, errorMessage);
 	}
 
-	public CreateGameResponse (String status, String message, int gameID)
+	public CreateGameResponse (String status, String message, long gameID)
 	{
 		super(status, message);  
 		this.gameID = gameID;
 	}
 
-	public int getGameID() {
+	public long getGameID() {
 		return gameID;
 	}
